Cancellation of outgoing requests in ethereum bridge
Current bridge realisation has some corner cases related to request approves:
- A user submits an outgoing request, peers collect approves. The user forgets about the signatures. Then >33% of peers changes. Now the signatures are invalid.
- Assume there are few pending outgoing requests and a peer that signed some of them. The peer gets removed. Now his signatures are invalid.

These cases can be solved by this way: a user submits approves to ethereum contract and gets an error because of invalid signatures. Then they submit a 'cancel' incoming request, which makes sure that the transaction has failed, but its input was correct. In this case, the request can be cancelled.

Signed-off-by: Vladislav Markushin <negigic@gmail.com>