Mark requests as done in ethereum bridge
Since we don't check that an outgoing request was finalized on the sidechain side, and thus client apps can't be sure that the request is done, we need a special function that will perform this check. `MarkAsDone` incoming request will check that the request with the given hash was used in the contract.

Signed-off-by: Vladislav Markushin <negigic@gmail.com>