[ETH Bridge] Ethereum XOR & VAL contracts peers synchronization
In the ethereum network we have 3 contracts: XOR, VAL and Bridge, each with the same set of peers. When adding/removing peer in that network, we should make sure to update the set on every contract. This is done by collecting additional proofs (in case of outgoing requests) by `OutgoingAdd(Remove)PeerCompat` and waiting for updates from all contracts (in case of incoming request) by `EthPeersSync`.

Signed-off-by: Vladislav Markushin <negigic@gmail.com>